add init_app
