+++ src/app.tsx
@@ -1,74 +1,11 @@
-import {
-    type CreateStepArgs,
-    type CreateStepStructure,
-    type MainStep,
-    type FlowSequence,
-    type SharedState,
-    Step,
-    PlanSelectOption,
-    AddonOption,
-} from './types';
-import { FlowStoreImpl as FlowStore } from './flowStore';
-import { NavigationProviderImpl as NavigationProvider } from './navigationProvider';
 import React from 'react';
-import { createPersonalInfoStep } from './ui/multi_step_form/personal_info/create';
-import { createSelectPlanStep } from './ui/multi_step_form/select_plan/create';
-import { createPlanAddonsStep } from './ui/multi_step_form/plan_addons/create';
-import { createSummaryStep } from './ui/multi_step_form/summary/create';
+import { type SharedState, PlanSelectOption, AddonOption } from './types';
 import { Dialog, Footer as DialogFooter } from '../src/ui/base/dialog/dialog';
-import type { FooterProps as DialogFooterProps } from '../src/ui/base/dialog/dialog';
 import { MultiStepFormHandlerImpl as MultiStepFormHandler } from './formHandler';
 import { PersonalInfoFormHandler } from './ui/multi_step_form/personal_info/formHandler';
 import { SelectPlanFormHandler } from './ui/multi_step_form/select_plan/formHandler';
 import { PlanAddonsFormHandler } from './ui/multi_step_form/plan_addons/formHandler';
-
-const steps: Record<MainStep, string> = {
-    personalInfo: 'Your info',
-    selectPlan: 'Select plan',
-    addons: 'add-ons',
-    summary: 'summary',
-};
-
-const sharedState: SharedState = {
-    selectPlan: {
-        type: 'monthly',
-        planType: 'arcade',
-        price: 9,
-    },
-    addons: {
-        type: 'monthly',
-        items: [],
-    },
-};
-
-const flowSequence: FlowSequence = {
-    personalInfo: {
-        subsequence: [],
-    },
-    selectPlan: {
-        subsequence: [],
-    },
-    addons: {
-        subsequence: [],
-    },
-    summary: {
-        subsequence: [Step.CONFIRMATION],
-    },
-};
-
-const flowStore = new FlowStore(
-    {
-        personalInfo: createPersonalInfoStep,
-        selectPlan: createSelectPlanStep,
-        addons: createPlanAddonsStep,
-        summary: createSummaryStep,
-        confirmation: createConfirmationStep,
-    },
-    Step.PERSONAL_INFO,
-    sharedState,
-    flowSequence
-);
-const navigationProvider = new NavigationProvider(flowStore);
+import { navigationProvider, flowStore, steps } from './init_app';
 
 function App() {
     const [rerenderCount, setRerenderCount] = React.useState(0);
@@ -125,26 +62,6 @@ function App() {
         multiStepFormHandler.load();
     }, []);
 
-    const Footer = ({ onNext, onBack, ...props }: DialogFooterProps) => {
-        const onNextHandler = () => {
-            onNext?.();
-            setRerenderCount(rerenderCount + 1);
-        };
-
-        const onBackHandler = () => {
-            onBack?.();
-            setRerenderCount(rerenderCount + 1);
-        };
-
-        return (
-            <DialogFooter
-                onBack={onBack && onBackHandler}
-                onNext={onNext && onNextHandler}
-                {...props}
-            />
-        );
-    };
-
     if (flowStore.createCurrentStep == null) {
         // TODO: if current step does not exist, close the dialog
         navigationProvider.close();
@@ -156,6 +73,20 @@ function App() {
         formHandler: multiStepFormHandler,
     });
 
+    const onBackHandler = () => {
+        stepStructure.onBack?.();
+        setRerenderCount(rerenderCount + 1);
+    };
+
+    const onNextHandler = () => {
+        stepStructure.onNext?.();
+        setRerenderCount(rerenderCount + 1);
+    };
+
+    const footer = (
+        <DialogFooter onBack={onBackHandler} onNext={onNextHandler} />
+    );
+
     return (
         <Dialog
             currentStep={step}
@@ -163,31 +94,9 @@ function App() {
             title={stepStructure.title}
             subtitle={stepStructure.subtitle}
             content={stepStructure.content}
-            footer={
-                <Footer
-                    disabledNext={!multiStepFormHandler.canSubmit(step)}
-                    onBack={stepStructure.onBack}
-                    onNext={stepStructure.onNext}
-                />
-            }
+            footer={footer}
         />
     );
-}
-
-function createConfirmationStep({
-    flowStore,
-    options: { sharedState },
-}: CreateStepArgs): CreateStepStructure<Step.SUMMARY> {
-    return ({ navigationProvider }) => {
-        return {
-            step: Step.SUMMARY,
-            structure: {
-                title: 'Confirmation',
-                subtitle: 'Confirmation subtitle',
-                content: <div>Content</div>,
-            },
-        };
-    };
 }
 
 export default App;
